Use <inttypes.h>
Why is it that everyone loves to redefine BYTE and WORD ?

+++ des.c
@@ -25,49 +25,49 @@
 #define SBOXBIT(a) (((a) & 0x20) | (((a) & 0x1f) >> 1) | (((a) & 0x01) << 4))
 
 /**************************** VARIABLES *****************************/
-static const BYTE sbox1[64] = {
+static const uint8_t sbox1[64] = {
 	14,  4,  13,  1,   2, 15,  11,  8,   3, 10,   6, 12,   5,  9,   0,  7,
 	 0, 15,   7,  4,  14,  2,  13,  1,  10,  6,  12, 11,   9,  5,   3,  8,
 	 4,  1,  14,  8,  13,  6,   2, 11,  15, 12,   9,  7,   3, 10,   5,  0,
 	15, 12,   8,  2,   4,  9,   1,  7,   5, 11,   3, 14,  10,  0,   6, 13
 };
-static const BYTE sbox2[64] = {
+static const uint8_t sbox2[64] = {
 	15,  1,   8, 14,   6, 11,   3,  4,   9,  7,   2, 13,  12,  0,   5, 10,
 	 3, 13,   4,  7,  15,  2,   8, 14,  12,  0,   1, 10,   6,  9,  11,  5,
 	 0, 14,   7, 11,  10,  4,  13,  1,   5,  8,  12,  6,   9,  3,   2, 15,
 	13,  8,  10,  1,   3, 15,   4,  2,  11,  6,   7, 12,   0,  5,  14,  9
 };
-static const BYTE sbox3[64] = {
+static const uint8_t sbox3[64] = {
 	10,  0,   9, 14,   6,  3,  15,  5,   1, 13,  12,  7,  11,  4,   2,  8,
 	13,  7,   0,  9,   3,  4,   6, 10,   2,  8,   5, 14,  12, 11,  15,  1,
 	13,  6,   4,  9,   8, 15,   3,  0,  11,  1,   2, 12,   5, 10,  14,  7,
 	 1, 10,  13,  0,   6,  9,   8,  7,   4, 15,  14,  3,  11,  5,   2, 12
 };
-static const BYTE sbox4[64] = {
+static const uint8_t sbox4[64] = {
 	 7, 13,  14,  3,   0,  6,   9, 10,   1,  2,   8,  5,  11, 12,   4, 15,
 	13,  8,  11,  5,   6, 15,   0,  3,   4,  7,   2, 12,   1, 10,  14,  9,
 	10,  6,   9,  0,  12, 11,   7, 13,  15,  1,   3, 14,   5,  2,   8,  4,
 	 3, 15,   0,  6,  10,  1,  13,  8,   9,  4,   5, 11,  12,  7,   2, 14
 };
-static const BYTE sbox5[64] = {
+static const uint8_t sbox5[64] = {
 	 2, 12,   4,  1,   7, 10,  11,  6,   8,  5,   3, 15,  13,  0,  14,  9,
 	14, 11,   2, 12,   4,  7,  13,  1,   5,  0,  15, 10,   3,  9,   8,  6,
 	 4,  2,   1, 11,  10, 13,   7,  8,  15,  9,  12,  5,   6,  3,   0, 14,
 	11,  8,  12,  7,   1, 14,   2, 13,   6, 15,   0,  9,  10,  4,   5,  3
 };
-static const BYTE sbox6[64] = {
+static const uint8_t sbox6[64] = {
 	12,  1,  10, 15,   9,  2,   6,  8,   0, 13,   3,  4,  14,  7,   5, 11,
 	10, 15,   4,  2,   7, 12,   9,  5,   6,  1,  13, 14,   0, 11,   3,  8,
 	 9, 14,  15,  5,   2,  8,  12,  3,   7,  0,   4, 10,   1, 13,  11,  6,
 	 4,  3,   2, 12,   9,  5,  15, 10,  11, 14,   1,  7,   6,  0,   8, 13
 };
-static const BYTE sbox7[64] = {
+static const uint8_t sbox7[64] = {
 	 4, 11,   2, 14,  15,  0,   8, 13,   3, 12,   9,  7,   5, 10,   6,  1,
 	13,  0,  11,  7,   4,  9,   1, 10,  14,  3,   5, 12,   2, 15,   8,  6,
 	 1,  4,  11, 13,  12,  3,   7, 14,  10, 15,   6,  8,   0,  5,   9,  2,
 	 6, 11,  13,  8,   1,  4,  10,  7,   9,  5,   0, 15,  14,  2,   3, 12
 };
-static const BYTE sbox8[64] = {
+static const uint8_t sbox8[64] = {
 	13,  2,   8,  4,   6, 15,  11,  1,  10,  9,   3, 14,   5,  0,  12,  7,
 	 1, 15,  13,  8,  10,  3,   7,  4,  12,  5,   6, 11,   0, 14,   9,  2,
 	 7, 11,   4,  1,   9, 12,  14,  2,   0,  6,  10, 13,  15,  3,   5,  8,
@@ -76,7 +76,7 @@ static const BYTE sbox8[64] = {
 
 /*********************** FUNCTION DEFINITIONS ***********************/
 // Initial (Inv)Permutation step
-void IP(WORD state[], const BYTE in[])
+void IP(uint32_t state[], const uint8_t in[])
 {
 	state[0] = BITNUM(in,57,31) | BITNUM(in,49,30) | BITNUM(in,41,29) | BITNUM(in,33,28) |
 				  BITNUM(in,25,27) | BITNUM(in,17,26) | BITNUM(in,9,25) | BITNUM(in,1,24) |
@@ -97,7 +97,7 @@ void IP(WORD state[], const BYTE in[])
 				  BITNUM(in,30,3) | BITNUM(in,22,2) | BITNUM(in,14,1) | BITNUM(in,6,0);
 }
 
-void InvIP(WORD state[], BYTE in[])
+void InvIP(uint32_t state[], uint8_t in[])
 {
 	in[0] = BITNUMINTR(state[1],7,7) | BITNUMINTR(state[0],7,6) | BITNUMINTR(state[1],15,5) |
 			  BITNUMINTR(state[0],15,4) | BITNUMINTR(state[1],23,3) | BITNUMINTR(state[0],23,2) |
@@ -132,10 +132,10 @@ void InvIP(WORD state[], BYTE in[])
 			  BITNUMINTR(state[1],24,1) | BITNUMINTR(state[0],24,0);
 }
 
-WORD f(WORD state, const BYTE key[])
+uint32_t f(uint32_t state, const uint8_t key[])
 {
-	BYTE lrgstate[6]; //,i;
-	WORD t1,t2;
+	uint8_t lrgstate[6]; //,i;
+	uint32_t t1,t2;
 
 	// Expantion Permutation
 	t1 = BITNUMINTL(state,31,0) | ((state & 0xf0000000) >> 1) | BITNUMINTL(state,4,5) |
@@ -190,15 +190,15 @@ WORD f(WORD state, const BYTE key[])
 	return(state);
 }
 
-void des_key_setup(const BYTE key[], BYTE schedule[][6], DES_MODE mode)
+void des_key_setup(const uint8_t key[], uint8_t schedule[][6], DES_MODE mode)
 {
-	WORD i, j, to_gen, C, D;
-	const WORD key_rnd_shift[16] = {1,1,2,2,2,2,2,2,1,2,2,2,2,2,2,1};
-	const WORD key_perm_c[28] = {56,48,40,32,24,16,8,0,57,49,41,33,25,17,
+	uint32_t i, j, to_gen, C, D;
+	const uint32_t key_rnd_shift[16] = {1,1,2,2,2,2,2,2,1,2,2,2,2,2,2,1};
+	const uint32_t key_perm_c[28] = {56,48,40,32,24,16,8,0,57,49,41,33,25,17,
 	                             9,1,58,50,42,34,26,18,10,2,59,51,43,35};
-	const WORD key_perm_d[28] = {62,54,46,38,30,22,14,6,61,53,45,37,29,21,
+	const uint32_t key_perm_d[28] = {62,54,46,38,30,22,14,6,61,53,45,37,29,21,
 	                             13,5,60,52,44,36,28,20,12,4,27,19,11,3};
-	const WORD key_compression[48] = {13,16,10,23,0,4,2,27,14,5,20,9,
+	const uint32_t key_compression[48] = {13,16,10,23,0,4,2,27,14,5,20,9,
 	                                  22,18,11,3,25,7,15,6,26,19,12,1,
 	                                  40,51,30,36,46,54,29,39,50,44,32,47,
 	                                  43,48,38,55,33,52,45,41,49,35,28,31};
@@ -230,9 +230,9 @@ void des_key_setup(const BYTE key[], BYTE schedule[][6], DES_MODE mode)
 	}
 }
 
-void des_crypt(const BYTE in[], BYTE out[], const BYTE key[][6])
+void des_crypt(const uint8_t in[], uint8_t out[], const uint8_t key[][6])
 {
-	WORD state[2],idx,t;
+	uint32_t state[2],idx,t;
 
 	IP(state,in);
 
@@ -247,7 +247,7 @@ void des_crypt(const BYTE in[], BYTE out[], const BYTE key[][6])
 	InvIP(state,out);
 }
 
-void three_des_key_setup(const BYTE key[], BYTE schedule[][16][6], DES_MODE mode)
+void three_des_key_setup(const uint8_t key[], uint8_t schedule[][16][6], DES_MODE mode)
 {
 	if (mode == DES_ENCRYPT) {
 		des_key_setup(&key[0],schedule[0],mode);
@@ -261,7 +261,7 @@ void three_des_key_setup(const BYTE key[], BYTE schedule[][16][6], DES_MODE mode
 	}
 }
 
-void three_des_crypt(const BYTE in[], BYTE out[], const BYTE key[][16][6])
+void three_des_crypt(const uint8_t in[], uint8_t out[], const uint8_t key[][16][6])
 {
 	des_crypt(in,out,key[0]);
 	des_crypt(out,out,key[1]);
